pull

+++ tests/training_tests/test_ubuntu_corpus_training.py
@@ -1,8 +1,7 @@
+from unittest.mock import Mock
 from io import BytesIO
-import unittest
 import tarfile
 import os
-from mock import Mock
 
 from tests.base_case import ChatBotTestCase
 from chatterbot.trainers import UbuntuCorpusTrainer
@@ -15,15 +14,15 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
 
     def setUp(self):
         super(UbuntuCorpusTrainerTestCase, self).setUp()
-        self.chatbot.set_trainer(UbuntuCorpusTrainer)
+        self.trainer = UbuntuCorpusTrainer(
+            self.chatbot,
+            show_training_progress=False
+        )
 
     def tearDown(self):
         super(UbuntuCorpusTrainerTestCase, self).tearDown()
-        import shutil
 
-        # Clean up by removing the corpus data directory
-        if os.path.exists(self.chatbot.trainer.data_directory):
-            shutil.rmtree(self.chatbot.trainer.data_directory)
+        self._remove_data()
 
     def _get_data(self):
 
@@ -43,22 +42,31 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
 
         return data1, data2
 
+    def _remove_data(self):
+        """
+        Clean up by removing the corpus data directory.
+        """
+        import shutil
+
+        if os.path.exists(self.trainer.data_directory):
+            shutil.rmtree(self.trainer.data_directory)
+
     def _create_test_corpus(self, data):
         """
         Create a small tar in a similar format to the
         Ubuntu corpus file in memory for testing.
         """
-        file_path = os.path.join(self.chatbot.trainer.data_directory, 'ubuntu_dialogs.tgz')
+        file_path = os.path.join(self.trainer.data_directory, 'ubuntu_dialogs.tgz')
         tar = tarfile.TarFile(file_path, 'w')
 
         tsv1 = BytesIO(data[0])
         tsv2 = BytesIO(data[1])
 
-        tarinfo = tarfile.TarInfo('ubuntu_dialogs/3/1.tsv')
+        tarinfo = tarfile.TarInfo('dialogs/3/1.tsv')
         tarinfo.size = len(data[0])
         tar.addfile(tarinfo, fileobj=tsv1)
 
-        tarinfo = tarfile.TarInfo('ubuntu_dialogs/3/2.tsv')
+        tarinfo = tarfile.TarInfo('dialogs/3/2.tsv')
         tarinfo.size = len(data[1])
         tar.addfile(tarinfo, fileobj=tsv2)
 
@@ -72,8 +80,10 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
         """
         Remove the test corpus file.
         """
-        file_path = os.path.join(self.chatbot.trainer.data_directory, 'ubuntu_dialogs.tgz')
-        os.remove(file_path)
+        file_path = os.path.join(self.trainer.data_directory, 'ubuntu_dialogs.tgz')
+
+        if os.path.exists(file_path):
+            os.remove(file_path)
 
     def _mock_get_response(self, *args, **kwargs):
         """
@@ -93,10 +103,10 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
 
         requests.get = Mock(side_effect=self._mock_get_response)
         download_url = 'https://example.com/download.tgz'
-        self.chatbot.trainer.download(download_url, show_status=False)
+        self.trainer.download(download_url, show_status=False)
 
         file_name = download_url.split('/')[-1]
-        downloaded_file_path = os.path.join(self.chatbot.trainer.data_directory, file_name)
+        downloaded_file_path = os.path.join(self.trainer.data_directory, file_name)
 
         requests.get.assert_called_with(download_url, stream=True)
         self.assertTrue(os.path.exists(downloaded_file_path))
@@ -110,12 +120,12 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
         """
         import requests
 
-        file_path = os.path.join(self.chatbot.trainer.data_directory, 'download.tgz')
+        file_path = os.path.join(self.trainer.data_directory, 'download.tgz')
         open(file_path, 'a').close()
 
         requests.get = Mock(side_effect=self._mock_get_response)
         download_url = 'https://example.com/download.tgz'
-        self.chatbot.trainer.download(download_url, show_status=False)
+        self.trainer.download(download_url, show_status=False)
 
         # Remove the dummy download_url
         os.remove(file_path)
@@ -126,33 +136,21 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
         """
         Test the case that the url being downloaded does not exist.
         """
-        raise unittest.SkipTest('This test needs to be created.')
+        self.skipTest('This test needs to be created.')
 
     def test_extract(self):
         """
         Test the extraction of text from a decompressed Ubuntu Corpus file.
         """
         file_object_path = self._create_test_corpus(self._get_data())
-        self.chatbot.trainer.extract(file_object_path)
+        self.trainer.extract(file_object_path)
 
         self._destroy_test_corpus()
-        corpus_path = os.path.join(self.chatbot.trainer.data_directory, 'ubuntu_dialogs', '3')
+        corpus_path = os.path.join(self.trainer.extracted_data_directory, 'dialogs', '3')
 
+        self.assertTrue(os.path.exists(self.trainer.extracted_data_directory))
         self.assertTrue(os.path.exists(os.path.join(corpus_path, '1.tsv')))
         self.assertTrue(os.path.exists(os.path.join(corpus_path, '2.tsv')))
-
-    def test_already_extracted(self):
-        """
-        Test that extraction is only done if the compressed file
-        has not already been extracted.
-        """
-        file_object_path = self._create_test_corpus(self._get_data())
-        created = self.chatbot.trainer.extract(file_object_path)
-        not_created = self.chatbot.trainer.extract(file_object_path)
-        self._destroy_test_corpus()
-
-        self.assertTrue(created)
-        self.assertFalse(not_created)
 
     def test_train(self):
         """
@@ -160,8 +158,29 @@ class UbuntuCorpusTrainerTestCase(ChatBotTestCase):
         """
         self._create_test_corpus(self._get_data())
 
-        self.chatbot.train()
+        self.trainer.train()
         self._destroy_test_corpus()
 
         response = self.chatbot.get_response('Is anyone there?')
         self.assertEqual(response, 'Yes')
+
+    def test_is_extracted(self):
+        """
+        Test that a check can be done for if the corpus has aleady been extracted.
+        """
+        file_object_path = self._create_test_corpus(self._get_data())
+        self.trainer.extract(file_object_path)
+
+        extracted = self.trainer.is_extracted(self.trainer.extracted_data_directory)
+        self._destroy_test_corpus()
+
+        self.assertTrue(extracted)
+
+    def test_is_not_extracted(self):
+        """
+        Test that a check can be done for if the corpus has aleady been extracted.
+        """
+        self._remove_data()
+        extracted = self.trainer.is_extracted(self.trainer.extracted_data_directory)
+
+        self.assertFalse(extracted)
